Setup phpbb_seo_test_case and some fix
